RESPSYSTEM/respwebapp/requirements.txt
absl-py==0.9.0
astroid==2.3.3
astunparse==1.6.3
audioread==2.1.8
bcrypt==3.1.7
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
DateTime==4.3
decorator==4.4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.14.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
numba==0.48.0
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.16
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0rc0
tensorflow-gpu==2.2.0rc3
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1
zope.interface==5.1.0